fix indices for courses and meal options

+++ packages/domain/src/core/course/useCases/deleteOne.ts
@@ -1,4 +1,5 @@
 import { Authorization, type DataSource, type Logger } from '../../..';
+import { type DBCourse } from '../../../data-source';
 import { type NanoId } from '../../shared';
 
 export interface DeleteOneCourseInput {
@@ -13,7 +14,27 @@ export async function deleteOne({ dataSourceAdapter, logger, context, request }:
 
     await Authorization.canMutateUserData({ context, dataSourceAdapter, logger, userId: cookId });
 
+    const course: DBCourse | undefined = await dataSourceAdapter.courseRepository.findOne({ cookId, menuId, courseId });
+
+    if (!course) {
+        logger.error('Tried to delete non existing course');
+        return false;
+    }
+
     const success: boolean = await dataSourceAdapter.courseRepository.deleteOne({ cookId, menuId, courseId });
+
+    if (success) {
+        await dataSourceAdapter.query(`
+            UPDATE
+                Courses
+            SET
+                Courses.index = Courses.index - 1
+            WHERE
+                Courses.cookId='${cookId}' AND
+                Courses.menuId='${menuId}' AND
+                Courses.index > ${course.index}
+        `);
+    }
 
     return success;
 }
