+++ packages/domain/src/core/meal-option/useCases/deleteOne.ts
@@ -1,4 +1,5 @@
 import { Authorization, type DataSource, type Logger } from '../../..';
+import { type DBMealOption } from '../../../data-source';
 import { type NanoId } from '../../shared';
 
 export interface DeleteOneMealOptionInput {
@@ -13,7 +14,28 @@ export async function deleteOne({ dataSourceAdapter, logger, context, request }:
 
     await Authorization.canMutateUserData({ context, dataSourceAdapter, logger, userId: cookId });
 
+    const mealOption: DBMealOption | undefined = await dataSourceAdapter.mealOptionRepository.findOne({ cookId, courseId, mealId });
+
+    if (!mealOption) {
+        logger.error('Tried to delete non existing meal option');
+        return false;
+    }
+
     const success: boolean = await dataSourceAdapter.mealOptionRepository.deleteOne({ cookId, courseId, mealId });
+
+    if (success) {
+        await dataSourceAdapter.query(`
+            UPDATE
+                MealOptions
+            SET
+                MealOptions.index = MealOptions.index - 1
+            WHERE
+                MealOptions.cookId='${cookId}' AND
+                MealOptions.courseId='${courseId}' AND
+                MealOptions.courseId='${courseId}' AND
+                MealOptions.index > ${mealOption.index}
+        `);
+    }
 
     return success;
 }
